requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
discord.py==1.5.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==1.0.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
yarl==1.5.1
